Performance improvement: update the gain value only if needed (#12)
GemvPropagationLossModel has been modified to store the gain value for each node pair. The gain value is stored in the map m_gainMap, together with the timestep at which the gain was computed. DoCalcRxPower checks if the gain value is already present in the map and if it needs to be updated (method NeedsUpdate ()). If so, the gain value is recomputed by looking at the gemv traces. If not, the value retrieved from the map is returned. To understand if a gain value needs to be updated, the method NeedsUpdate () compares the timestep saved in the map with the current timestep.