Add the possibility to install a fake RAN AI (#26)
This PR adds the possibility to install a fake RAN AI. Instead of forwarding the collected metrics to the python interface, they are saved in a text file "RanAiStats.txt". When using this feature, no action is taken by the kitti APPs, i.e., the application mode remains always the same.
The parameter useFakeRanAi to enable/disable the fake RAN AI was added to the gem-ai.cc script.